Change cursor from bigint to string
Working on the upcoming support for tasks backed by custom enumerators
surfaced the need for non-numeric cursors.

For instance, one might have a task that iterates over resources backed
by some API providing its own opaque string cursor, which would be
impossible to serialize to a bigint column.

As such, this changes the maintenance_tasks_runs.cursor column to be a
string. For this to work, we must deserialize the cursor back to an
integer in the Array and CSV collection cases, while the ActiveRecord
case is handled transparently.

Co-authored-by: Adrianna Chang <adrianna.chang@shopify.com>
Co-authored-by: Sam Bostock <sam.bostock@shopify.com>